+++ tests/shell/test_qemu_launch_stage3_vm.sh
@@ -20,10 +20,22 @@ assert_contains() {
   [[ "${haystack}" == *"${needle}"* ]] || fail "expected to find '${needle}' in '${haystack}'"
 }
 
+assert_file_contains() {
+  local path="$1"
+  local needle="$2"
+  grep -Fq "${needle}" "${path}" || fail "expected '${needle}' in ${path}"
+}
+
 assert_equals() {
   local actual="$1"
   local expected="$2"
   [[ "${actual}" == "${expected}" ]] || fail "expected '${expected}', got '${actual}'"
+}
+
+write_fixture_file() {
+  local path="$1"
+  local content="${2:-fixture}"
+  printf '%s\n' "${content}" > "${path}"
 }
 
 mark_stage3_launch_globals_used() {
@@ -33,6 +45,7 @@ mark_stage3_launch_globals_used() {
     "${QEMU_SMP-}" \
     "${QEMU_MEMORY_MIB-}" \
     "${QEMU_BOOT_STRICT-}" \
+    "${QEMU_ATTACH_HOST_DISKS-}" \
     "${QEMU_LAUNCH_DRY_RUN-}" \
     "${QEMU_DAEMONIZE-}" \
     "${QEMU_DISPLAY_MODE-}" \
@@ -98,6 +111,7 @@ reset_launcher_state() {
   QEMU_BOOTDISK_ID='bootdisk'
   QEMU_BOOTDISK_MODEL='virtio-blk-pci'
   QEMU_BOOTDISK_BOOTINDEX='1'
+  QEMU_ATTACH_HOST_DISKS='1'
   QEMU_LAUNCH_DRY_RUN='1'
   QEMU_DAEMONIZE='1'
   QEMU_DISPLAY_MODE='none'
@@ -166,12 +180,12 @@ test_build_qemu_cmd_uses_boot_disk_and_tcp_serial() {
   RPOOL_DISK1="${temp_dir}/rpool1.img"
   QEMU_SERIAL_MODE='tcp'
   : > "${QCOW_IMAGE}"
-  : > "${EFI_FIRM}"
-  : > "${EFI_VARS_TEMPLATE}"
-  : > "${BPOOL_DISK0}"
-  : > "${BPOOL_DISK1}"
-  : > "${RPOOL_DISK0}"
-  : > "${RPOOL_DISK1}"
+  write_fixture_file "${EFI_FIRM}" OVMF
+  write_fixture_file "${EFI_VARS_TEMPLATE}" VARS
+  write_fixture_file "${BPOOL_DISK0}"
+  write_fixture_file "${BPOOL_DISK1}"
+  write_fixture_file "${RPOOL_DISK0}"
+  write_fixture_file "${RPOOL_DISK1}"
 
   build_qemu_cmd
   rendered="${QEMU_CMD[*]}"
@@ -187,6 +201,14 @@ test_build_qemu_cmd_uses_boot_disk_and_tcp_serial() {
   rm -rf "${temp_dir}"
 }
 
+test_setup_launcher_logging_degrades_without_dev_fd() {
+  reset_launcher_state
+  LAUNCHER_LOG_ENABLE='1'
+  LAUNCHER_LOG_INITIALIZED=0
+
+  assert_file_contains "${LAUNCH_SCRIPT}" 'Launcher log file requested but /dev/fd is unavailable'
+}
+
 test_build_qemu_cmd_supports_pty_serial() {
   local temp_dir rendered
   temp_dir="$(mktemp -d)"
@@ -200,12 +222,12 @@ test_build_qemu_cmd_supports_pty_serial() {
   RPOOL_DISK1="${temp_dir}/rpool1.img"
   QEMU_SERIAL_MODE='pty'
   : > "${QCOW_IMAGE}"
-  : > "${EFI_FIRM}"
-  : > "${EFI_VARS_TEMPLATE}"
-  : > "${BPOOL_DISK0}"
-  : > "${BPOOL_DISK1}"
-  : > "${RPOOL_DISK0}"
-  : > "${RPOOL_DISK1}"
+  write_fixture_file "${EFI_FIRM}" OVMF
+  write_fixture_file "${EFI_VARS_TEMPLATE}" VARS
+  write_fixture_file "${BPOOL_DISK0}"
+  write_fixture_file "${BPOOL_DISK1}"
+  write_fixture_file "${RPOOL_DISK0}"
+  write_fixture_file "${RPOOL_DISK1}"
 
   build_qemu_cmd
   rendered="${QEMU_CMD[*]}"
@@ -226,12 +248,12 @@ test_build_qemu_cmd_supports_target_disk_boot() {
   RPOOL_DISK0="${temp_dir}/rpool0.img"
   RPOOL_DISK1="${temp_dir}/rpool1.img"
   QEMU_BOOT_SOURCE='target-disks'
-  : > "${EFI_FIRM}"
-  : > "${EFI_VARS_TEMPLATE}"
-  : > "${BPOOL_DISK0}"
-  : > "${BPOOL_DISK1}"
-  : > "${RPOOL_DISK0}"
-  : > "${RPOOL_DISK1}"
+  write_fixture_file "${EFI_FIRM}" OVMF
+  write_fixture_file "${EFI_VARS_TEMPLATE}" VARS
+  write_fixture_file "${BPOOL_DISK0}"
+  write_fixture_file "${BPOOL_DISK1}"
+  write_fixture_file "${RPOOL_DISK0}"
+  write_fixture_file "${RPOOL_DISK1}"
 
   build_qemu_cmd
   rendered="${QEMU_CMD[*]}"
@@ -240,6 +262,29 @@ test_build_qemu_cmd_supports_target_disk_boot() {
   assert_contains "${rendered}" 'ide-hd,drive=bpool0,bus=ahci.1,serial=bpool-0,bootindex=1'
   assert_contains "${rendered}" 'ide-hd,drive=bpool1,bus=ahci.2,serial=bpool-1'
   assert_contains "${rendered}" '-boot strict=on'
+  rm -rf "${temp_dir}"
+}
+
+test_build_qemu_cmd_can_skip_host_disks_for_qcow_boot() {
+  local temp_dir rendered
+  temp_dir="$(mktemp -d)"
+
+  reset_launcher_state
+  QCOW_IMAGE="${temp_dir}/vm.qcow2"
+  EFI_FIRM="${temp_dir}/OVMF_CODE.fd"
+  QEMU_ATTACH_HOST_DISKS='0'
+  write_fixture_file "${QCOW_IMAGE}"
+  write_fixture_file "${EFI_FIRM}" OVMF
+  write_fixture_file "${EFI_VARS_TEMPLATE}" VARS
+
+  build_qemu_cmd
+  rendered="${QEMU_CMD[*]}"
+
+  assert_contains "${rendered}" "if=none,id=${QEMU_BOOTDISK_ID},file=${QCOW_IMAGE},format=qcow2"
+  [[ "${rendered}" != *'id=bpool0'* ]] || fail 'optional host disks should not attach bpool0'
+  [[ "${rendered}" != *'id=bpool1'* ]] || fail 'optional host disks should not attach bpool1'
+  [[ "${rendered}" != *'id=rpool0'* ]] || fail 'optional host disks should not attach rpool0'
+  [[ "${rendered}" != *'id=rpool1'* ]] || fail 'optional host disks should not attach rpool1'
   rm -rf "${temp_dir}"
 }
 
@@ -256,12 +301,12 @@ test_build_qemu_cmd_supports_alias_network_mode() {
   RPOOL_DISK1="${temp_dir}/rpool1.img"
   QEMU_NETWORK_MODE='alias'
   : > "${QCOW_IMAGE}"
-  : > "${EFI_FIRM}"
-  : > "${EFI_VARS_TEMPLATE}"
-  : > "${BPOOL_DISK0}"
-  : > "${BPOOL_DISK1}"
-  : > "${RPOOL_DISK0}"
-  : > "${RPOOL_DISK1}"
+  write_fixture_file "${EFI_FIRM}" OVMF
+  write_fixture_file "${EFI_VARS_TEMPLATE}" VARS
+  write_fixture_file "${BPOOL_DISK0}"
+  write_fixture_file "${BPOOL_DISK1}"
+  write_fixture_file "${RPOOL_DISK0}"
+  write_fixture_file "${RPOOL_DISK1}"
 
   build_qemu_cmd
   rendered="${QEMU_CMD[*]}"
@@ -287,14 +332,14 @@ test_build_qemu_cmd_supports_memory_drive_manifest() {
 {"drives":[{"path":"${temp_dir}/mem0.qcow2","format":"qcow2","serial":"mem-portage-cache","device_model":"virtio-blk-pci"}]}
 EOF
   QEMU_MEMORY_DRIVES_FILE="${manifest}"
-  : > "${QCOW_IMAGE}"
-  : > "${EFI_FIRM}"
-  : > "${EFI_VARS_TEMPLATE}"
-  : > "${BPOOL_DISK0}"
-  : > "${BPOOL_DISK1}"
-  : > "${RPOOL_DISK0}"
-  : > "${RPOOL_DISK1}"
-  : > "${temp_dir}/mem0.qcow2"
+  write_fixture_file "${QCOW_IMAGE}"
+  write_fixture_file "${EFI_FIRM}" OVMF
+  write_fixture_file "${EFI_VARS_TEMPLATE}" VARS
+  write_fixture_file "${BPOOL_DISK0}"
+  write_fixture_file "${BPOOL_DISK1}"
+  write_fixture_file "${RPOOL_DISK0}"
+  write_fixture_file "${RPOOL_DISK1}"
+  write_fixture_file "${temp_dir}/mem0.qcow2"
 
   build_qemu_cmd
   rendered="${QEMU_CMD[*]}"
@@ -319,6 +364,76 @@ test_validate_boot_source_rejects_invalid_value() {
   assert_contains "${output}" 'Unsupported QEMU_BOOT_SOURCE'
 }
 
+test_validate_qcow_image_allows_missing_qcow_in_dry_run() {
+  local temp_dir output status
+  temp_dir="$(mktemp -d)"
+
+  reset_launcher_state
+  QCOW_IMAGE="${temp_dir}/missing.qcow2"
+  QEMU_LAUNCH_DRY_RUN='1'
+
+  set +e
+  output="$(validate_qcow_image 2>&1)"
+  status=$?
+  set -e
+
+  assert_equals "${status}" '0'
+  assert_contains "${output}" 'QCOW_IMAGE is missing during dry-run'
+  rm -rf "${temp_dir}"
+}
+
+test_validate_qcow_image_requires_qcow_outside_dry_run() {
+  local temp_dir output status
+  temp_dir="$(mktemp -d)"
+
+  reset_launcher_state
+  QCOW_IMAGE="${temp_dir}/missing.qcow2"
+  QEMU_LAUNCH_DRY_RUN='0'
+
+  set +e
+  output="$(validate_qcow_image 2>&1)"
+  status=$?
+  set -e
+
+  assert_equals "${status}" '1'
+  assert_contains "${output}" 'QCOW_IMAGE is missing'
+  rm -rf "${temp_dir}"
+}
+
+test_validate_host_disks_skips_optional_disks_for_qcow_boot() {
+  local temp_dir
+  temp_dir="$(mktemp -d)"
+
+  reset_launcher_state
+  BPOOL_DISK0="${temp_dir}/missing-bpool0.img"
+  BPOOL_DISK1="${temp_dir}/missing-bpool1.img"
+  RPOOL_DISK0="${temp_dir}/missing-rpool0.img"
+  RPOOL_DISK1="${temp_dir}/missing-rpool1.img"
+  QEMU_ATTACH_HOST_DISKS='0'
+
+  validate_host_disks
+  rm -rf "${temp_dir}"
+}
+
+test_validate_host_disks_requires_target_disks_even_when_optional_disabled() {
+  local temp_dir output status
+  temp_dir="$(mktemp -d)"
+
+  reset_launcher_state
+  BPOOL_DISK0="${temp_dir}/missing-bpool0.img"
+  QEMU_BOOT_SOURCE='target-disks'
+  QEMU_ATTACH_HOST_DISKS='0'
+
+  set +e
+  output="$(validate_host_disks 2>&1)"
+  status=$?
+  set -e
+
+  assert_equals "${status}" '1'
+  assert_contains "${output}" 'BPOOL_DISK0 is missing'
+  rm -rf "${temp_dir}"
+}
+
 test_main_dry_run_prints_stage3_vm_command() {
   local temp_dir output status
   temp_dir="$(mktemp -d)"
@@ -330,13 +445,13 @@ test_main_dry_run_prints_stage3_vm_command() {
   BPOOL_DISK1="${temp_dir}/bpool1.img"
   RPOOL_DISK0="${temp_dir}/rpool0.img"
   RPOOL_DISK1="${temp_dir}/rpool1.img"
-  : > "${QCOW_IMAGE}"
-  : > "${EFI_FIRM}"
-  : > "${EFI_VARS_TEMPLATE}"
-  : > "${BPOOL_DISK0}"
-  : > "${BPOOL_DISK1}"
-  : > "${RPOOL_DISK0}"
-  : > "${RPOOL_DISK1}"
+  write_fixture_file "${QCOW_IMAGE}"
+  write_fixture_file "${EFI_FIRM}" OVMF
+  write_fixture_file "${EFI_VARS_TEMPLATE}" VARS
+  write_fixture_file "${BPOOL_DISK0}"
+  write_fixture_file "${BPOOL_DISK1}"
+  write_fixture_file "${RPOOL_DISK0}"
+  write_fixture_file "${RPOOL_DISK1}"
 
   set +e
   output="$(main 2>&1)"
@@ -363,12 +478,12 @@ test_main_dry_run_prints_target_disk_boot_plan() {
   RPOOL_DISK0="${temp_dir}/rpool0.img"
   RPOOL_DISK1="${temp_dir}/rpool1.img"
   QEMU_BOOT_SOURCE='target-disks'
-  : > "${EFI_FIRM}"
-  : > "${EFI_VARS_TEMPLATE}"
-  : > "${BPOOL_DISK0}"
-  : > "${BPOOL_DISK1}"
-  : > "${RPOOL_DISK0}"
-  : > "${RPOOL_DISK1}"
+  write_fixture_file "${EFI_FIRM}" OVMF
+  write_fixture_file "${EFI_VARS_TEMPLATE}" VARS
+  write_fixture_file "${BPOOL_DISK0}"
+  write_fixture_file "${BPOOL_DISK1}"
+  write_fixture_file "${RPOOL_DISK0}"
+  write_fixture_file "${RPOOL_DISK1}"
 
   set +e
   output="$(main 2>&1)"
@@ -425,13 +540,13 @@ test_build_qemu_cmd_supports_spice_qxl_and_agent() {
   QEMU_SPICE_PORT='5931'
   QEMU_SPICE_ADDRESS='127.0.0.1'
   QEMU_SPICE_OPTIONS='port=5931,addr=127.0.0.1,disable-ticketing=on'
-  : > "${QCOW_IMAGE}"
-  : > "${EFI_FIRM}"
-  : > "${EFI_VARS_TEMPLATE}"
-  : > "${BPOOL_DISK0}"
-  : > "${BPOOL_DISK1}"
-  : > "${RPOOL_DISK0}"
-  : > "${RPOOL_DISK1}"
+  write_fixture_file "${QCOW_IMAGE}"
+  write_fixture_file "${EFI_FIRM}" OVMF
+  write_fixture_file "${EFI_VARS_TEMPLATE}" VARS
+  write_fixture_file "${BPOOL_DISK0}"
+  write_fixture_file "${BPOOL_DISK1}"
+  write_fixture_file "${RPOOL_DISK0}"
+  write_fixture_file "${RPOOL_DISK1}"
 
   build_qemu_cmd
   rendered="${QEMU_CMD[*]}"
@@ -446,14 +561,20 @@ test_build_qemu_cmd_supports_spice_qxl_and_agent() {
 
 test_default_launcher_log_file_uses_requested_format
 test_build_qemu_cmd_uses_boot_disk_and_tcp_serial
+test_setup_launcher_logging_degrades_without_dev_fd
 test_build_qemu_cmd_supports_pty_serial
 test_build_qemu_cmd_supports_target_disk_boot
+test_build_qemu_cmd_can_skip_host_disks_for_qcow_boot
 test_build_qemu_cmd_supports_alias_network_mode
 test_build_qemu_cmd_supports_memory_drive_manifest
 test_video_device_defaults_to_qxl_for_spice
 test_validate_display_backend_rejects_missing_spice
 test_build_qemu_cmd_supports_spice_qxl_and_agent
 test_validate_boot_source_rejects_invalid_value
+test_validate_qcow_image_allows_missing_qcow_in_dry_run
+test_validate_qcow_image_requires_qcow_outside_dry_run
+test_validate_host_disks_skips_optional_disks_for_qcow_boot
+test_validate_host_disks_requires_target_disks_even_when_optional_disabled
 test_main_dry_run_prints_stage3_vm_command
 test_main_dry_run_prints_target_disk_boot_plan
